summarizer/requirements.txt
pip install \
  requests \
  beautifulsoup4 \
  python-dotenv \
  openai \
  tabulate \
